+++ lib/src/ui/failure_messages.dart
@@ -45,6 +45,7 @@ String describeFailure(AppLocalizations l10n, Object error) {
     FailureCode.proxyFormat => l10n.failureProxyFormat,
     FailureCode.proxyPort => l10n.failureProxyPort,
     FailureCode.initializationFailed => l10n.failureInitializationFailed(detail),
+    FailureCode.voiceBankClearFailed => l10n.failureVoiceBankClearFailed(detail),
   };
 }